refactor(BreadcrumbEllipsis): remove the interface to index file
