--- src/components/navigation/BreadcrumbEllipsis/interfaces/BreadcrumbEllipsis.Route.interface.ts
@@ -1,5 +1,0 @@
-export interface IRoute {
-  label: string;
-  path: string;
-  id: string;
-}
