--- src/components/navigation/BreadcrumbEllipsis/interfaces/BreadcrumbEllipsis.interface.ts
@@ -1,7 +1,0 @@
-import { Typos } from "../types/BreadcrumbEllipsis.Typos.type";
-import { IRoute } from "./BreadcrumbEllipsis.Route.interface";
-
-export interface IBreadcrumbEllipsisProps {
-  size?: Typos;
-  routes: IRoute[];
-}
